Move attribute checker to it's own file
Both attribute utility and attribute checker (pre-expansion) were mixed
within the same file/translation unit. Splitting them apart will allow
a shorter header to include and make both file easier to navigate.

gcc/rust/ChangeLog:

	* Make-lang.in: Add a new rust-attribute-checker file.
	* checks/errors/rust-builtin-attribute-checker.h (check_valid_attribute_for_item):
	Change BuiltinAttributeChecker doc comment.
	* rust-session-manager.cc: Include the new rust-attribute-checker
	header.
	* util/rust-attributes.cc (AttributeChecker::AttributeChecker): Move to
	rust-attribute-checker.cc.
	(AttributeChecker::go): Likewise.
	(is_proc_macro_type): Likewise.
	(check_proc_macro_non_function): Likewise.
	(check_proc_macro_non_root): Likewise.
	(AttributeChecker::visit): Likewise.
	* util/rust-attributes.h (class AttributeChecker): Move to
	rust-attribute-checker.h.
	* checks/errors/rust-attribute-checker.cc: New file.
	* checks/errors/rust-attribute-checker.h: New file.
